♻️  (#449) Commet Editor 코드 리팩터링
